requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-bootstrap5==21.3
django-excel-response==2.0.5
django-filter==22.1
django-import-export==2.8.0
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
idna==3.3
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==2.4.11
Pillow==9.2.0
PyYAML==6.0
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
<<<<<<< HEAD
=======
urllib3==1.26.12
xlrd==2.0.1
>>>>>>> 1b8f61522213f19984cd4b90fbbb980b04f1b597
xlwt==1.3.0
